Remove knex wrapper
