+++ modules/database/server-ts/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@gqlapp/core-common": "^0.1.0"
+    "@gqlapp/core-common": "^0.1.0",
+    "glob": "^7.1.3"
   }
 }
