+++ packages/server/package.json
@@ -17,9 +17,9 @@
     "eslint": "eslint --fix --ext js --ext jsx --ext json src",
     "tslint": "tslint --fix -p tsconfig.json -c ../../tslint.json",
     "lint": "yarn eslint && yarn tslint",
-    "seed": "yarn migrate && node ../../tools/knex seed:run",
-    "migrate": "node ../../tools/knex migrate:latest",
-    "rollback": "node ../../tools/knex migrate:rollback",
+    "seed": "yarn migrate && knex seed:run",
+    "migrate": "knex migrate:latest",
+    "rollback": "knex migrate:rollback",
     "watch": "zen watch",
     "heroku-postbuild": "yarn build && yarn seed && yarn stripe:setup",
     "cli": "node ../../tools/cli",
@@ -158,7 +158,6 @@
     "chai-http": "^4.0.0",
     "cross-env": "^5.1.1",
     "eslint": "^4.13.1",
-    "glob": "^7.1.3",
     "ip": "^1.1.5",
     "jsdom": "^11.5.1",
     "mkdirp": "^0.5.1",
